Add duration to time conversion for tasks
The taskview and edittaskviews now display a human friendly format for
the duration as a string. This has been included in both the
MainActivity in the TaskViews as well as a summarytext for the edit and
add task activities.

+++ app/src/main/java/com/sudosays/tempo/activities/AddTask.kt
@@ -36,7 +36,7 @@ class AddTask : AppCompatActivity() {
     {
         if ((taskEditview.nameEditText.text.isNotEmpty())&&(taskEditview.durationEditText.text.isNotEmpty())) {
             val last_position = sharedPrefs.getInt("last_position",0)
-            val task = Task(0, taskEditview.nameEditText.text.toString(), taskEditview.durationEditText.text.toString().toFloat(), last_position)
+            val task = Task(0, taskEditview.nameEditText.text.toString(), taskEditview.durationEditText.text.toString().toInt(), last_position)
             TaskInsertAsync(db).execute(task)
 
             with(sharedPrefs.edit()) {
